change install condition for nvim, update install logic for nvim, kitty, zen

+++ tools/nvim.sh
@@ -1,16 +1,15 @@
 #!/bin/bash
-which nvim && exit 1
+[[ -d $HOME/.local/nvim-linux64/ ]] && exit 1
 
 url="https://github.com/neovim/neovim/releases/download/
     v0.9.5/nvim-linux64.tar.gz"
 curl -L ${url//[[:space:]]/} -o hic || exit 1
 
-mkdir -p $HOME/.local/bin/
-
-rm -rf $HOME/.local/nvim-linux64/
+mkdir -p $HOME/.local/
 tar xf hic -C $HOME/.local/
 rm hic
 
+mkdir -p $HOME/.local/bin/
 rm -f $HOME/.local/bin/nvim
 ln -s $HOME/.local/{nvim-linux64,}/bin/nvim
